feat: Show auth type in document history
Surface when a history event was performed via the API or MCP, e.g.
"Tom published via MCP". Moves AuthenticationType to shared types and
exposes authType on the event presenter.
