+++ app/scenes/Document/components/History/EventListItem.tsx
@@ -13,6 +13,7 @@ import { useTranslation } from "react-i18next";
 import styled, { css } from "styled-components";
 import { s } from "@shared/styles";
 import Text from "@shared/components/Text";
+import { AuthenticationType } from "@shared/types";
 import type Document from "~/models/Document";
 import type Event from "~/models/Event";
 import Time from "~/components/Time";
@@ -21,6 +22,13 @@ import Logger from "~/utils/Logger";
 type Props = {
   document: Document;
   item: Event<Document>;
+};
+
+/** Authentication types that are surfaced in history, mapped to a display name. */
+const authTypeNames: Partial<Record<AuthenticationType, string>> = {
+  [AuthenticationType.API]: "API",
+  [AuthenticationType.OAUTH]: "API",
+  [AuthenticationType.MCP]: "MCP",
 };
 
 const EventListItem = ({ item }: Props) => {
@@ -88,14 +96,19 @@ const EventListItem = ({ item }: Props) => {
     return null;
   }
 
+  const authTypeName = item.authType ? authTypeNames[item.authType] : undefined;
+
   return (
     <EventItem>
       <IconWrapper size="xsmall" type="secondary">
         {icon}
       </IconWrapper>
       <Text size="xsmall" type="secondary">
-        {meta} &middot;{" "}
-        <Time dateTime={item.createdAt} relative shorten addSuffix />
+        {meta}
+        {authTypeName
+          ? ` ${t("via {{ authType }}", { authType: authTypeName })}`
+          : ""}{" "}
+        &middot; <Time dateTime={item.createdAt} relative shorten addSuffix />
       </Text>
     </EventItem>
   );
